+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.5.0 (unreleased)
 ------------------
 
+- #2408 Support DX type catalogs lookup
+- #2407 Fix analyses sort order in Transposed Multi Results Form
+- #2406 Fix missing interim fields in Transposed Multi Results Form
 - #2400 Add Transposed Multi Results Form
 - #2402 Fix user cannot enter future date for DateSampled when sampling enabled
 - #2401 Fix OverflowError when calculating datetime.min date for left-hand TZs